requirements.txt
Flask==3.0.2
python-docx==1.1.2
gunicorn==21.2.0
Werkzeug==3.0.1 